Split test_company_views into different modules
