+++ datahub/company/test/test_company_views_v4.py
@@ -3,13 +3,9 @@ from datetime import datetime
 
 import factory
 import pytest
-import reversion
-from django.utils.timezone import now
 from freezegun import freeze_time
 from rest_framework import status
 from rest_framework.reverse import reverse
-from rest_framework.settings import api_settings
-from reversion.models import Version
 
 from datahub.company.constants import BusinessTypeConstant
 from datahub.company.models import Company, OneListTier
@@ -18,11 +14,8 @@ from datahub.company.test.factories import (
     AdviserFactory,
     CompaniesHouseCompanyFactory,
     CompanyFactory,
-    DuplicateCompanyFactory,
-    OneListCoreTeamMemberFactory,
 )
 from datahub.core.constants import Country, EmployeeRange, HeadquarterType, TurnoverRange, UKRegion
-from datahub.core.reversion import EXCLUDED_BASE_MODEL_FIELDS
 from datahub.core.test_utils import (
     APITestMixin,
     create_test_user,
@@ -1878,491 +1871,3 @@ class TestAddCompany(APITestMixin):
 
         assert response.status_code == status.HTTP_400_BAD_REQUEST
         assert response.json() == expected_error
-
-
-class TestArchiveCompany(APITestMixin):
-    """Archive company tests."""
-
-    def test_archive_company_no_reason(self):
-        """Test company archive."""
-        company = CompanyFactory()
-        url = reverse('api-v4:company:archive', kwargs={'pk': company.id})
-        response = self.api_client.post(url)
-
-        assert response.status_code == status.HTTP_400_BAD_REQUEST
-        assert response.json() == {
-            'reason': ['This field is required.'],
-        }
-
-    def test_archive_company_reason(self):
-        """Test company archive."""
-        company = CompanyFactory()
-        url = reverse('api-v4:company:archive', kwargs={'pk': company.id})
-        response = self.api_client.post(url, data={'reason': 'foo'})
-
-        assert response.status_code == status.HTTP_200_OK
-        response_data = response.json()
-        assert response_data['archived']
-        assert response_data['archived_reason'] == 'foo'
-        assert response_data['id'] == str(company.id)
-
-    def test_archive_company_invalid_address(self):
-        """
-        Test archiving a company when the company has an invalid trading address and missing
-        UK region.
-        """
-        company = CompanyFactory(
-            registered_address_country_id=Country.united_kingdom.value.id,
-            trading_address_town='',
-            uk_region_id=None,
-        )
-        url = reverse('api-v4:company:archive', kwargs={'pk': company.id})
-        response = self.api_client.post(url, data={'reason': 'foo'})
-
-        assert response.status_code == status.HTTP_200_OK
-        response_data = response.json()
-        assert response_data['archived']
-        assert response_data['archived_reason'] == 'foo'
-
-
-class TestUnarchiveCompany(APITestMixin):
-    """Unarchive company tests."""
-
-    def test_unarchive_company_invalid_address(self):
-        """
-        Test unarchiving a company when the company has an invalid trading address and missing
-        UK region.
-        """
-        company = CompanyFactory(
-            address_country_id=Country.united_kingdom.value.id,
-            trading_address_town='',
-            uk_region_id=None,
-            archived=True,
-            archived_reason='Dissolved',
-        )
-        url = reverse('api-v4:company:unarchive', kwargs={'pk': company.id})
-        response = self.api_client.post(url)
-
-        assert response.status_code == status.HTTP_200_OK
-        assert not response.json()['archived']
-
-    def test_unarchive_company(self):
-        """Unarchive a company."""
-        company = CompanyFactory(
-            archived=True,
-            archived_on=now(),
-            archived_reason='foo',
-        )
-        url = reverse('api-v4:company:unarchive', kwargs={'pk': company.id})
-        response = self.api_client.post(url)
-
-        assert response.status_code == status.HTTP_200_OK
-        response_data = response.json()
-        assert not response_data['archived']
-        assert response_data['archived_reason'] == ''
-        assert response_data['id'] == str(company.id)
-
-    def test_cannot_unarchive_duplicate_company(self):
-        """Test that a duplicate company cannot be unarchived."""
-        company = DuplicateCompanyFactory()
-        url = reverse('api-v4:company:unarchive', kwargs={'pk': company.id})
-        response = self.api_client.post(url)
-
-        assert response.status_code == status.HTTP_400_BAD_REQUEST
-        assert response.json() == {
-            api_settings.NON_FIELD_ERRORS_KEY:
-                [
-                    'This record is no longer in use and its data has been transferred to another '
-                    'record for the following reason: Duplicate record.',
-                ],
-        }
-
-
-class TestCompanyVersioning(APITestMixin):
-    """
-    Tests for versions created when interacting with the company endpoints.
-    """
-
-    def test_add_creates_a_new_version(self):
-        """Test that creating a company creates a new version."""
-        assert Version.objects.count() == 0
-
-        response = self.api_client.post(
-            reverse('api-v4:company:collection'),
-            data={
-                'name': 'Acme',
-                'trading_name': 'Trading name',
-                'business_type': {'id': BusinessTypeConstant.company.value.id},
-                'sector': {'id': random_obj_for_model(Sector).id},
-                'address': {
-                    'line_1': '75 Stramford Road',
-                    'town': 'London',
-                    'country': {
-                        'id': Country.united_kingdom.value.id,
-                    },
-                },
-                'uk_region': {'id': UKRegion.england.value.id},
-            },
-        )
-
-        assert response.status_code == status.HTTP_201_CREATED
-        response_data = response.json()
-        assert response_data['name'] == 'Acme'
-        assert response_data['trading_name'] == 'Trading name'
-
-        company = Company.objects.get(pk=response_data['id'])
-
-        # check version created
-        assert Version.objects.get_for_object(company).count() == 1
-        version = Version.objects.get_for_object(company).first()
-        assert version.revision.user == self.user
-        assert version.field_dict['name'] == 'Acme'
-        assert version.field_dict['trading_names'] == ['Trading name']
-        assert not any(set(version.field_dict) & set(EXCLUDED_BASE_MODEL_FIELDS))
-
-    def test_promoting_a_ch_company_creates_a_new_version(self):
-        """Test that promoting a CH company to full company creates a new version."""
-        assert Version.objects.count() == 0
-        CompaniesHouseCompanyFactory(company_number=1234567890)
-
-        response = self.api_client.post(
-            reverse('api-v4:company:collection'),
-            data={
-                'name': 'Acme',
-                'company_number': 1234567890,
-                'business_type': BusinessTypeConstant.company.value.id,
-                'sector': random_obj_for_model(Sector).id,
-                'address': {
-                    'line_1': '75 Stramford Road',
-                    'town': 'London',
-                    'country': {
-                        'id': Country.united_kingdom.value.id,
-                    },
-                },
-                'uk_region': UKRegion.england.value.id,
-            },
-        )
-
-        assert response.status_code == status.HTTP_201_CREATED
-
-        company = Company.objects.get(pk=response.json()['id'])
-
-        # check version created
-        assert Version.objects.get_for_object(company).count() == 1
-        version = Version.objects.get_for_object(company).first()
-        assert version.field_dict['company_number'] == '1234567890'
-
-    def test_add_400_doesnt_create_a_new_version(self):
-        """Test that if the endpoint returns 400, no version is created."""
-        assert Version.objects.count() == 0
-
-        response = self.api_client.post(
-            reverse('api-v4:company:collection'),
-            data={'name': 'Acme'},
-        )
-
-        assert response.status_code == status.HTTP_400_BAD_REQUEST
-        assert Version.objects.count() == 0
-
-    def test_update_creates_a_new_version(self):
-        """Test that updating a company creates a new version."""
-        company = CompanyFactory(name='Foo ltd.')
-
-        assert Version.objects.get_for_object(company).count() == 0
-
-        response = self.api_client.patch(
-            reverse('api-v4:company:item', kwargs={'pk': company.pk}),
-            data={'name': 'Acme'},
-        )
-
-        assert response.status_code == status.HTTP_200_OK
-        assert response.json()['name'] == 'Acme'
-
-        # check version created
-        assert Version.objects.get_for_object(company).count() == 1
-        version = Version.objects.get_for_object(company).first()
-        assert version.revision.user == self.user
-        assert version.field_dict['name'] == 'Acme'
-
-    def test_update_400_doesnt_create_a_new_version(self):
-        """Test that if the endpoint returns 400, no version is created."""
-        company = CompanyFactory()
-
-        response = self.api_client.patch(
-            reverse('api-v4:company:item', kwargs={'pk': company.pk}),
-            data={'trading_name': 'a' * 600},
-        )
-
-        assert response.status_code == status.HTTP_400_BAD_REQUEST
-        assert Version.objects.get_for_object(company).count() == 0
-
-    def test_archive_creates_a_new_version(self):
-        """Test that archiving a company creates a new version."""
-        company = CompanyFactory()
-        assert Version.objects.get_for_object(company).count() == 0
-
-        url = reverse('api-v4:company:archive', kwargs={'pk': company.id})
-        response = self.api_client.post(url, data={'reason': 'foo'})
-
-        assert response.status_code == status.HTTP_200_OK
-        response_data = response.json()
-        assert response_data['archived']
-        assert response_data['archived_reason'] == 'foo'
-
-        # check version created
-        assert Version.objects.get_for_object(company).count() == 1
-        version = Version.objects.get_for_object(company).first()
-        assert version.revision.user == self.user
-        assert version.field_dict['archived']
-        assert version.field_dict['archived_reason'] == 'foo'
-
-    def test_archive_400_doesnt_create_a_new_version(self):
-        """Test that if the endpoint returns 400, no version is created."""
-        company = CompanyFactory()
-        assert Version.objects.get_for_object(company).count() == 0
-
-        url = reverse('api-v4:company:archive', kwargs={'pk': company.id})
-        response = self.api_client.post(url)
-
-        assert response.status_code == status.HTTP_400_BAD_REQUEST
-        assert Version.objects.get_for_object(company).count() == 0
-
-    def test_unarchive_creates_a_new_version(self):
-        """Test that unarchiving a company creates a new version."""
-        company = CompanyFactory(
-            archived=True, archived_on=now(), archived_reason='foo',
-        )
-        assert Version.objects.get_for_object(company).count() == 0
-
-        url = reverse('api-v4:company:unarchive', kwargs={'pk': company.id})
-        response = self.api_client.post(url)
-
-        assert response.status_code == status.HTTP_200_OK
-        response_data = response.json()
-        assert not response_data['archived']
-        assert response_data['archived_reason'] == ''
-
-        # check version created
-        assert Version.objects.get_for_object(company).count() == 1
-        version = Version.objects.get_for_object(company).first()
-        assert version.revision.user == self.user
-        assert not version.field_dict['archived']
-
-
-class TestAuditLogView(APITestMixin):
-    """Tests for the audit log view."""
-
-    def test_audit_log_view(self):
-        """Test retrieval of audit log."""
-        initial_datetime = now()
-        with reversion.create_revision():
-            company = CompanyFactory(
-                description='Initial desc',
-            )
-
-            reversion.set_comment('Initial')
-            reversion.set_date_created(initial_datetime)
-            reversion.set_user(self.user)
-
-        changed_datetime = now()
-        with reversion.create_revision():
-            company.description = 'New desc'
-            company.save()
-
-            reversion.set_comment('Changed')
-            reversion.set_date_created(changed_datetime)
-            reversion.set_user(self.user)
-
-        versions = Version.objects.get_for_object(company)
-        version_id = versions[0].id
-        url = reverse('api-v4:company:audit-item', kwargs={'pk': company.pk})
-
-        response = self.api_client.get(url)
-        response_data = response.json()['results']
-
-        # No need to test the whole response
-        assert len(response_data) == 1
-        entry = response_data[0]
-
-        assert entry['id'] == version_id
-        assert entry['user']['name'] == self.user.name
-        assert entry['comment'] == 'Changed'
-        assert entry['timestamp'] == format_date_or_datetime(changed_datetime)
-        assert entry['changes']['description'] == ['Initial desc', 'New desc']
-        assert not set(EXCLUDED_BASE_MODEL_FIELDS) & entry['changes'].keys()
-
-
-class TestOneListGroupCoreTeam(APITestMixin):
-    """Tests for getting the One List Core Team of a company's group."""
-
-    @pytest.mark.parametrize(
-        'build_company',
-        (
-            # as subsidiary
-            lambda: CompanyFactory(
-                global_headquarters=CompanyFactory(one_list_account_owner=None),
-            ),
-            # as single company
-            lambda: CompanyFactory(
-                global_headquarters=None,
-                one_list_account_owner=None,
-            ),
-        ),
-        ids=('as_subsidiary', 'as_non_subsidiary'),
-    )
-    def test_empty_list(self, build_company):
-        """
-        Test that if there's no Global Account Manager and no Core Team
-        member for a company's Global Headquarters, the endpoint returns
-        an empty list.
-        """
-        company = build_company()
-
-        url = reverse(
-            'api-v4:company:one-list-group-core-team',
-            kwargs={'pk': company.pk},
-        )
-        response = self.api_client.get(url)
-
-        assert response.status_code == status.HTTP_200_OK
-        assert response.json() == []
-
-    @pytest.mark.parametrize(
-        'build_company',
-        (
-            # as subsidiary
-            lambda gam: CompanyFactory(
-                global_headquarters=CompanyFactory(one_list_account_owner=gam),
-            ),
-            # as single company
-            lambda gam: CompanyFactory(
-                global_headquarters=None,
-                one_list_account_owner=gam,
-            ),
-        ),
-        ids=('as_subsidiary', 'as_non_subsidiary'),
-    )
-    def test_with_only_global_account_manager(self, build_company):
-        """
-        Test that if there is a Global Account Manager but no Core Team
-        member for a company's Global Headquarters, the endpoint returns
-        a list with only that adviser in it.
-        """
-        global_account_manager = AdviserFactory()
-        company = build_company(global_account_manager)
-
-        url = reverse(
-            'api-v4:company:one-list-group-core-team',
-            kwargs={'pk': company.pk},
-        )
-        response = self.api_client.get(url)
-
-        assert response.status_code == status.HTTP_200_OK
-        assert response.json() == [
-            {
-                'adviser': {
-                    'id': str(global_account_manager.pk),
-                    'name': global_account_manager.name,
-                    'first_name': global_account_manager.first_name,
-                    'last_name': global_account_manager.last_name,
-                    'dit_team': {
-                        'id': str(global_account_manager.dit_team.pk),
-                        'name': global_account_manager.dit_team.name,
-                        'uk_region': {
-                            'id': str(global_account_manager.dit_team.uk_region.pk),
-                            'name': global_account_manager.dit_team.uk_region.name,
-                        },
-                        'country': {
-                            'id': str(global_account_manager.dit_team.country.pk),
-                            'name': global_account_manager.dit_team.country.name,
-                        },
-                    },
-                },
-                'is_global_account_manager': True,
-            },
-        ]
-
-    @pytest.mark.parametrize(
-        'build_company',
-        (
-            # as subsidiary
-            lambda gam: CompanyFactory(
-                global_headquarters=CompanyFactory(one_list_account_owner=gam),
-            ),
-            # as single company
-            lambda gam: CompanyFactory(
-                global_headquarters=None,
-                one_list_account_owner=gam,
-            ),
-        ),
-        ids=('as_subsidiary', 'as_non_subsidiary'),
-    )
-    @pytest.mark.parametrize(
-        'with_global_account_manager',
-        (True, False),
-        ids=lambda val: f'{"With" if val else "Without"} global account manager',
-    )
-    def test_with_core_team_members(self, build_company, with_global_account_manager):
-        """
-        Test that if there are Core Team members for a company's Global Headquarters,
-        the endpoint returns a list with these advisers in it.
-        """
-        team_member_advisers = AdviserFactory.create_batch(
-            3,
-            first_name=factory.Iterator(
-                ('Adam', 'Barbara', 'Chris'),
-            ),
-        )
-        global_account_manager = team_member_advisers[0] if with_global_account_manager else None
-
-        company = build_company(global_account_manager)
-        group_global_headquarters = company.global_headquarters or company
-        OneListCoreTeamMemberFactory.create_batch(
-            len(team_member_advisers),
-            company=group_global_headquarters,
-            adviser=factory.Iterator(team_member_advisers),
-        )
-
-        url = reverse(
-            'api-v4:company:one-list-group-core-team',
-            kwargs={'pk': company.pk},
-        )
-        response = self.api_client.get(url)
-
-        assert response.status_code == status.HTTP_200_OK
-        assert response.json() == [
-            {
-                'adviser': {
-                    'id': str(adviser.pk),
-                    'name': adviser.name,
-                    'first_name': adviser.first_name,
-                    'last_name': adviser.last_name,
-                    'dit_team': {
-                        'id': str(adviser.dit_team.pk),
-                        'name': adviser.dit_team.name,
-                        'uk_region': {
-                            'id': str(adviser.dit_team.uk_region.pk),
-                            'name': adviser.dit_team.uk_region.name,
-                        },
-                        'country': {
-                            'id': str(adviser.dit_team.country.pk),
-                            'name': adviser.dit_team.country.name,
-                        },
-                    },
-                },
-                'is_global_account_manager': adviser is global_account_manager,
-            }
-            for adviser in team_member_advisers
-        ]
-
-    def test_404_with_invalid_company(self):
-        """
-        Test that if the company doesn't exist, the endpoint returns 404.
-        """
-        url = reverse(
-            'api-v4:company:one-list-group-core-team',
-            kwargs={'pk': '00000000-0000-0000-0000-000000000000'},
-        )
-        response = self.api_client.get(url)
-
-        assert response.status_code == status.HTTP_404_NOT_FOUND
