+++ datahub/company/test/test_company_views_v4_archiving.py
@@ -1,0 +1,106 @@
+from django.utils.timezone import now
+from rest_framework import status
+from rest_framework.reverse import reverse
+from rest_framework.settings import api_settings
+
+from datahub.company.test.factories import CompanyFactory, DuplicateCompanyFactory
+from datahub.core.constants import Country
+from datahub.core.test_utils import APITestMixin
+
+
+class TestArchiveCompany(APITestMixin):
+    """Archive company tests."""
+
+    def test_archive_company_no_reason(self):
+        """Test company archive."""
+        company = CompanyFactory()
+        url = reverse('api-v4:company:archive', kwargs={'pk': company.id})
+        response = self.api_client.post(url)
+
+        assert response.status_code == status.HTTP_400_BAD_REQUEST
+        assert response.json() == {
+            'reason': ['This field is required.'],
+        }
+
+    def test_archive_company_reason(self):
+        """Test company archive."""
+        company = CompanyFactory()
+        url = reverse('api-v4:company:archive', kwargs={'pk': company.id})
+        response = self.api_client.post(url, data={'reason': 'foo'})
+
+        assert response.status_code == status.HTTP_200_OK
+        response_data = response.json()
+        assert response_data['archived']
+        assert response_data['archived_reason'] == 'foo'
+        assert response_data['id'] == str(company.id)
+
+    def test_archive_company_invalid_address(self):
+        """
+        Test archiving a company when the company has an invalid trading address and missing
+        UK region.
+        """
+        company = CompanyFactory(
+            registered_address_country_id=Country.united_kingdom.value.id,
+            trading_address_town='',
+            uk_region_id=None,
+        )
+        url = reverse('api-v4:company:archive', kwargs={'pk': company.id})
+        response = self.api_client.post(url, data={'reason': 'foo'})
+
+        assert response.status_code == status.HTTP_200_OK
+        response_data = response.json()
+        assert response_data['archived']
+        assert response_data['archived_reason'] == 'foo'
+
+
+class TestUnarchiveCompany(APITestMixin):
+    """Unarchive company tests."""
+
+    def test_unarchive_company_invalid_address(self):
+        """
+        Test unarchiving a company when the company has an invalid trading address and missing
+        UK region.
+        """
+        company = CompanyFactory(
+            address_country_id=Country.united_kingdom.value.id,
+            trading_address_town='',
+            uk_region_id=None,
+            archived=True,
+            archived_reason='Dissolved',
+        )
+        url = reverse('api-v4:company:unarchive', kwargs={'pk': company.id})
+        response = self.api_client.post(url)
+
+        assert response.status_code == status.HTTP_200_OK
+        assert not response.json()['archived']
+
+    def test_unarchive_company(self):
+        """Unarchive a company."""
+        company = CompanyFactory(
+            archived=True,
+            archived_on=now(),
+            archived_reason='foo',
+        )
+        url = reverse('api-v4:company:unarchive', kwargs={'pk': company.id})
+        response = self.api_client.post(url)
+
+        assert response.status_code == status.HTTP_200_OK
+        response_data = response.json()
+        assert not response_data['archived']
+        assert response_data['archived_reason'] == ''
+        assert response_data['id'] == str(company.id)
+
+    def test_cannot_unarchive_duplicate_company(self):
+        """Test that a duplicate company cannot be unarchived."""
+        company = DuplicateCompanyFactory()
+        url = reverse('api-v4:company:unarchive', kwargs={'pk': company.id})
+        response = self.api_client.post(url)
+
+        assert response.status_code == status.HTTP_400_BAD_REQUEST
+        assert response.json() == {
+            api_settings.NON_FIELD_ERRORS_KEY:
+                [
+                    'This record is no longer in use and its data has been transferred to another '
+                    'record for the following reason: Duplicate record.',
+                ],
+        }
